requirements.txt
django 
psycopg2 
djangorestframework 
django-filter 
drf-yasg 
python-dotenv
